% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_validator_functions.R
\name{build_linking_df}
\alias{build_linking_df}
\title{Build linking df}
\usage{
build_linking_df(tibble_list)
}
\arguments{
\item{tibble_list}{the list of tibbles outputted by \code{load_spreadsheet}}
}
\value{
a tibble with fields type_module_1, type_module_2 and the fields in common
between them
}
\description{
\code{build_linking_df} builds a dataframe from the result of running
\code{get_linking_fields} over a multisheet spreadsheet. The wrangler can
then review whether all requisite linking fields are present and use as input
to other functions to help validation.
}
