% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_validator_functions.R
\name{check_uploaded_files}
\alias{check_uploaded_files}
\title{Check uploaded files s3}
\usage{
check_uploaded_files(ssheet_file_names, s3_file_names)
}
\arguments{
\item{ssheet_file_names}{result of}

\item{s3_file_names}{result of `list_s3_files$file_name`}
}
\description{
\code{check_uploaded_files} cross references the files listed in the
spreadsheet against the files currently in the s3 and warns the user if there are
inconsistencies.
}
