% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_interacting_functions.R
\name{copy_s3}
\alias{copy_s3}
\title{Copy s3 files between HCA s3 buckets}
\usage{
copy_s3(source_s3, dest_s3, user)
}
\arguments{
\item{{source_s3}}{the url of the s3 where the files to be copied are}

\item{{dest_s3}}{the url of the s3 where you want the files to go}

\item{{user}}{your username}
}
\description{
\code{copy_s3} copies files between two HCA s3 buckets. It will not work for
buckets outside the HCA
}
