% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{count_accessions}
\alias{count_accessions}
\title{Count accessions}
\usage{
count_accessions(acc_string)
}
\arguments{
\item{acc_string}{a string}
}
\value{
an integer
}
\description{
\code{count_accessions} counts the number of accessions in a \code{||}
delimited list. (Not exported)
}
