% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{count_bundles}
\alias{count_bundles}
\title{Count bundles}
\usage{
count_bundles(tibble_list)
}
\arguments{
\item{tibble_list}{a list of tibbles outputted by \code{load_spreadsheet}}
}
\description{
\code{count_bundles} estimates the number of bundles based on the number of
levels in the `process_id` field in the `Sequence file` tab of the
spreadsheet.
}
