% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_interacting_functions.R
\name{create_s3}
\alias{create_s3}
\title{Create s3 upload area}
\usage{
create_s3(
  project_shorthand,
  hca_util_path = paste0(reticulate::virtualenv_root(), "/ranglr_env/bin/hca-util"),
  permissions = NA,
  profile = NA
)
}
\arguments{
\item{project_shorthand}{the `shortname of the dataset e.g. my-cool-project}

\item{hca_util_path}{the path to installed hca-util}

\item{permissions}{allowed actions (permissions) on new area. u for upload, x
for delete and d for download. Default is ux}

\item{profile}{specify name of admin profile, otherwise uses default}
}
\value{
the uuid of the s3 bucket
}
\description{
\code{create_s3} uses the hca-util tool to create an s3 upload area for a
particular project.
}
\details{
Assumes you have installed and configured the \code{setup_hca_util} command.
Can optionally provide the path to where you have installed hca-util
}
