% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_api_functions.R
\name{get_audr_info}
\alias{get_audr_info}
\title{Query ingest api for project information to add to an AUDR ticket}
\usage{
get_audr_info(submission_environment = "prod", project_uuid)
}
\arguments{
\item{submission_environment}{either `staging` or `prod`}

\item{project_uuid}{the uuid of the project you want info about}
}
\description{
\code{get_audr_info} provides an easy to use wrapper to query the
ingest api using a project uuid to get required info for the AUDR ticket
}
