% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_authors_from_contributors}
\alias{get_authors_from_contributors}
\title{Get author list from contributors}
\usage{
get_authors_from_contributors(tibble_list)
}
\arguments{
\item{tibble_list}{the list of tibbles outputted from `load_spreadsheet()`}
}
\value{
a string of the authors from the contributor tab
}
\description{
\code{get_author_from contributors} formats the publication authors based on what is in the contributors tab and formats in the correct format, initials, surname and \code{||} separated.
}
