% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_col_names}
\alias{get_col_names}
\title{Get column names per sheet}
\usage{
get_col_names(tibble_list)
}
\arguments{
\item{tibble_list}{list of tibbles outputted by \code{load_spreadsheet}}
}
\value{
a tibble with two columns, \code{sheet_name} and \code{col_name}
}
\description{
\code{get_col_names} gets all the column names for a sheet and returns a
tibble linking the column names to the name of the sheet.
}
