% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_field_name}
\alias{get_field_name}
\title{Get field name}
\usage{
get_field_name(long_name)
}
\arguments{
\item{long_name}{a string}
}
\value{
field_name
}
\description{
\code{get_field_name} gets the specific field name in the long `.`` delimited
programmatic name. It splits the field by the `.` and returns the last string.
(not exported)
}
