% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_file_functions.R
\name{get_file_format}
\alias{get_file_format}
\title{Get file format}
\usage{
get_file_format(file_name)
}
\arguments{
\item{file_name}{the string filename}
}
\description{
\code{get_file_format} attempts to recognise the file format of the file by
looking for common fastq patterns.
}
\details{
currently searches for \code{fq.gz | fastq.gz}
}
