% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_file_name_fields}
\alias{get_file_name_fields}
\title{Get filename fields in a spreadsheet}
\usage{
get_file_name_fields(tibble_list)
}
\arguments{
\item{tibble_list}{a list of tibbles outputted from \code{load_spreadsheet}}
}
\value{
a list of filenames
}
\description{
\code{get_file_name_fields} gets all the filenames listed in a spreadsheet so
that they can a wrangler can check if they have been uploaded to the
appropriate s3 bucket for a project.
}
