% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_validator_functions.R
\name{get_hca_ontology}
\alias{get_hca_ontology}
\title{Read current HCA ontology}
\usage{
get_hca_ontology(gh_repo = "https://github.com/HumanCellAtlas/ontology")
}
\description{
\code{get_hca_ontology} reads the current dcp ontologies from github and
returns a dataframe. By default reads from github but can optionally supply
a file path if up to date obo's already downloaded locally.
}
