% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_file_functions.R
\name{get_lane_index}
\alias{get_lane_index}
\title{Get lane index}
\usage{
get_lane_index(file_name)
}
\arguments{
\item{file_name}{the string filename}
}
\description{
\code{get_lane_index} attempts to recognise the lane index of a fastq file by
 recognising the lane index of the file.
}
\details{
The regex used to recognise the lane is: \code{[-_]L\\d+}
}
