% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_linking_fields}
\alias{get_linking_fields}
\title{Get linking fields}
\usage{
get_linking_fields(sheet_1, sheet_2)
}
\arguments{
\item{sheet_1}{a tibble, first sheet to compare}

\item{sheet_2}{a tibble, second sheet to compare}
}
\value{
a tibble with columns \code{type_module_1}, \code{type_module_2},
\code{linking_field}
}
\description{
\code{get_linking_fields} takes a metadata spreadsheet list and determines
which fields are used as linking between them.
}
