% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_not_field_name}
\alias{get_not_field_name}
\title{Get everything except field name}
\usage{
get_not_field_name(long_name)
}
\arguments{
\item{long_name}{a string}
}
\value{
not the field name
}
\description{
\code{get_not_field_name} gets the long `.` delimited programmatic name.apart
from the final field. It splits the field by the `.` and returns everything
except the last string.
}
