% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_ontologies}
\alias{get_ontologies}
\title{Get ontologies}
\usage{
get_ontologies(tibble_list)
}
\arguments{
\item{tibble_list}{the list of tibbles outputted from `load_spreadsheet()`}
}
\value{
a dataframe with sheet_name, field_name, text, ontology,
ontology_label
}
\description{
\code{get_ontologies} goes through the spreadsheet to mine out all the
ontology terms so that a wrangler can quickly look at the list of ontologies
for a sanity check.
}
