% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_file_functions.R
\name{get_read_index}
\alias{get_read_index}
\title{Get read index}
\usage{
get_read_index(file_name)
}
\arguments{
\item{file_name}{the string filename}
}
\description{
\code{get_read_index} attempts to recognise the read index of a fastq file in
a string and returns the string required by hca metadata standard, i.e. one of
index1, index2, read1, read2.
}
\details{
Currently recognised patterns are: R1, R2, I1, I2
}
