% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_rows}
\alias{get_rows}
\title{Get rows}
\usage{
get_rows(tibble_list, sheet_name, field_name)
}
\arguments{
\item{tibble_list}{the loaded metadata spreadsheet}

\item{sheet_name}{the name of sheet}

\item{field_name}{a list of the programmatic names of the desired columns}
}
\value{
a tibble with the `field_names` given
}
\description{
\code{get_rows} gets a set of rows from the spreadsheet by the sheet name and
field name.
}
