% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_api_functions.R
\name{get_submission_env_from_uuid}
\alias{get_submission_env_from_uuid}
\title{Query ingest api to get submission envelope id from submission uuid}
\usage{
get_submission_env_from_uuid(sub_uuid, sub_environment = c("staging", "prod"))
}
\arguments{
\item{submission_uuid}{the uuid for the submission}

\item{submission_environment}{the environment the submission is in, one of "staging or "prod"}
}
\description{
Query ingest api to get submission envelope id from submission uuid
}
