% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_type_module}
\alias{get_type_module}
\title{Get type module}
\usage{
get_type_module(sheet)
}
\arguments{
\item{sheet}{a tibble from a tibble_list}
}
\value{
the sheet name as a string
}
\description{
\code{get_type_module} infers the sheet's name by getting the first part of
the type and module fields from the first column name. Assumes that the first
column is a typical column name for the spreadsheet.
}
