% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{load_spreadsheet}
\alias{load_spreadsheet}
\title{load a metadata spreadsheet}
\usage{
load_spreadsheet(path)
}
\arguments{
\item{path}{a string.}
}
\value{
a list of tibbles for each sheet in the spreadsheet.
}
\description{
load a metadata spreadsheet
}
\examples{
load_spreadsheet("~/name_of_spreadsheet.xlsx")
}
