% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pull_field}
\alias{pull_field}
\title{Pull field}
\usage{
pull_field(sheet_name, field_name, tibble_list)
}
\arguments{
\item{sheet_name}{a string}

\item{field_name}{programmatic field name}

\item{tibble_list}{the list of tibbles outputted from \code{load_spreadsheet}}
}
\value{
a vector of strings the length of the field
}
\description{
\code{pull_field} pulls a particular field from the spreadsheet based on the
sheet name and the field name
}
