% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pull_field_levels}
\alias{pull_field_levels}
\title{Pull field levels}
\usage{
pull_field_levels(sheet_name, field_name, tibble_list)
}
\arguments{
\item{sheet_name}{a string}

\item{field_name}{programmatic field name}

\item{tibble_list}{the list of tibbles outputted from \code{load_spreadsheet}}
}
\value{
a vector of unique values for a field
}
\description{
\code{pull_field_levels} pulls a list of unique levels for a particular field
}
