% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_api_functions.R
\name{query_existing_submission}
\alias{query_existing_submission}
\title{Query ingest api for uuids in an existing submission}
\usage{
query_existing_submission(submission_uuid,
  submission_environment = c("staging", "prod"),
  entity_type = c("biomaterials", "files", "processes", "projects",
  "protocols"))
}
\arguments{
\item{submission_uuid}{a string}

\item{submission_environment}{either `staging` or `prod`}

\item{entity_type}{one of processes, projects, protocols, files, biomaterials}
}
\description{
\code{query_exising_submission} provides an easy to use wrapper to query the
ingest api using a submission envelope uuid
}
