% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_validator_functions.R
\name{query_hca_ontology_term}
\alias{query_hca_ontology_term}
\title{Query HCA ontology by term}
\usage{
query_hca_ontology_term(ontology_term, ontology,
  base_url = "https://ontology.staging.data.humancellatlas.org/api/ontologies")
}
\arguments{
\item{ontology_term}{a vector of string ontology terms e.g. EFO:0009737}

\item{ontology}{the name of the ontology to query}

\item{base_url}{the base url by default set to \code{https://ontology.staging.data.humancellatlas.org/api/ontologies}}
}
\value{
a dataframe with api results
}
\description{
\code{query_hca_ontology_term} queries the current hca ontology using the
hcao ols API.
}
