% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_validator_functions.R
\name{query_ontology_text}
\alias{query_ontology_text}
\title{Query OLS for free text}
\usage{
query_ontology_text(free_text,
  base_url = "https://ontology.staging.data.humancellatlas.org/api/search?q=")
}
\arguments{
\item{free_text}{to use for the query}

\item{base_url}{by default uses the hcao staging API}
}
\value{
a tibble with query results
}
\description{
\code{query_ontology_text} allows the user to query the OLS with free text
using the `search` endpoint
}
