% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_api_functions.R
\name{query_zoom}
\alias{query_zoom}
\title{Query text in zoom}
\usage{
query_zoom(free_text, ontology = NULL)
}
\arguments{
\item{free_text}{the text to search}
}
\value{
tibble of results
}
\description{
\code{query_zoom} searches the zoom API for a free text ontology field and
returns possible likely ontologies.
}
