% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranglr.R
\docType{package}
\name{ranglr}
\alias{ranglr}
\alias{ranglr-package}
\title{ranglr: a set of tools for HCA Data Wranglers}
\description{
A collection of tools to help with Wrangling and wrangling review
process. The aim is to automate some processes and improve efficiency and
productivity.
}
\section{ranglr functions}{

...
}

