% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{split_field_list}
\alias{split_field_list}
\title{Split list of fields}
\usage{
split_field_list(field_list)
}
\arguments{
\item{field_list}{a || delimited string}
}
\value{
a vector of strings
}
\description{
\code{split_field_list} takes as input a single string that is made up of
multiple fields that are delimited by `||`` and returns a string vector
}
