% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_querying_functions.R
\name{ssh_ec2}
\alias{ssh_ec2}
\title{Connect to EC2}
\usage{
ssh_ec2(username, ec2_url = "tool.staging.data.humancellatlas.org")
}
\arguments{
\item{username}{string}

\item{ec2_url}{ec2 url, `tool.staging.data.humancellatlas.org` by default}
}
\value{
connection ssh connection object
}
\description{
\code{ssh_ec2} uses ssh to connect to the EC2 to be able to access s3 buckets.
}
\details{
The user must already have permissions to access the EC2 already set up.
}
