% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarise_spreadsheet}
\alias{summarise_spreadsheet}
\title{Summarise a spreadsheet}
\usage{
summarise_spreadsheet(tibble_list)
}
\arguments{
\item{tibble_list}{a list of tibbles outputted by \code{load_spreadsheet}}
}
\value{
a tibble with summary counts
}
\description{
\code{summarise_spreadsheet} provides a useful summary of the data that is
currently in a spreadsheet so that a wrangler can get a quick and easy
overview. This function prints a nicely formatted table as well as returning
the tibble from which the table was derived.
}
