% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_interacting_functions.R
\name{sync_s3}
\alias{sync_s3}
\title{Sync files between hca-util and ingest upload areas}
\usage{
sync_s3(
  source_s3,
  target_s3,
  profile = NA,
  hca_util_path = paste0(reticulate::virtualenv_root(), "/ranglr_env/bin/hca-util")
)
}
\arguments{
\item{source_s3}{the uuid of the hca-util upload area to copy from}

\item{target_s3}{the full address of the ingest submission upload area}

\item{profile}{specify name of admin profile, otherwise uses default}

\item{hca_util_path}{the path to installed hca-util}
}
\description{
\code{sync_s3} copies files from an hca-util upload area into an ingest
submission upload area
}
