% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tally_ontology}
\alias{tally_ontology}
\title{Tally ontology use}
\usage{
tally_ontology(tibble_list, ontology_field, sheet_name)
}
\arguments{
\item{tibble_list}{list of tibbles from load_spreadsheet}

\item{ontology_field}{the name of the field up to the last bit, a.k.a not field}

\item{sheet_name}{name of the sheet/tab where the ontology is found}
}
\value{
a tibble with fields and counts of occurrences in the spreadsheet
}
\description{
\code{tally_ontology} counts the occurrences of a particular ontology within
a spreadsheet
}
