% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_validator_functions.R
\name{validate_link_levels}
\alias{validate_link_levels}
\title{Validate link levels}
\usage{
validate_link_levels(linking_df, tibble_list)
}
\arguments{
\item{linking_df}{output from running \code{build_linking_df}}

\item{tibble_list}{the loaded spreadsheet from \code{load_spreadsheet}}
}
\value{
linking_df with two additional columns
}
\description{
\code{validate_link_levels} checks the levels of a linking field are
consistent between two tabs of the spreadsheet. It is currently not very
elegantly written.
}
