% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_interacting_functions.R
\name{create_s3}
\alias{create_s3}
\title{Create s3 upload area}
\usage{
create_s3(user, project_shorthand,
  upload_api_key = Sys.getenv("STAGING_UPLOAD_API_KEY"),
  upload_env_name = "staging", submission_n = 0)
}
\arguments{
\item{user}{your username}

\item{project_shorthand}{the `-` delimited name of the dataset e.g. my-cool-project}

\item{upload_api_key}{the upload api string}

\item{upload_env_name}{the name of the environment, i.e. staging, dev,
integration}

\item{submission_n}{by default 0 but can be changed if it is a bucket for
a subsequent submission for the same project}
}
\value{
the url of the s3 bucket of the form \code{s3://org-humancellatlas-upload-staging/aaaaaaaa-bbbb-cccc-dddd-e012345f6789/}
}
\description{
\code{create_s3} creates an s3 upload area for a particular project. By
assumes you have the staging api key saved in your .Renviron file and called
"STAGING_UPLOAD_API_KEY"
}
\examples{
create_s3("FirstInitialLastName", "my-cool-project")
}
