% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expand_rows}
\alias{expand_rows}
\title{Expand rows by field level}
\usage{
expand_rows(split_tibble, tibble_list)
}
\arguments{
\item{split_tibble}{a tibble with consistent field levels}

\item{tibble_list}{the tibble imported with \code{load_spreadsheet}}
}
\value{
a tibble with expanded rows
}
\description{
\code{expand_rows} takes a tibble with a consistent number of nested field
levels and expands (unnests) the rows.
}
