% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_file_names}
\alias{get_file_names}
\title{Get filenames in the spreadsheet}
\usage{
get_file_names(tibble_list, field = NULL)
}
\arguments{
\item{tibble_list}{outputted from \code{load_spreadsheet}}
}
\value{
a vector of string filenames
}
\description{
\code{get_file_names} gets a list of the filenames in the spreadsheet. By
default returns all filenames in the spreadsheet.
}
