% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_interacting_functions.R
\name{list_s3_files}
\alias{list_s3_files}
\title{List files in S3}
\usage{
list_s3_files(s3_url, user)
}
\arguments{
\item{s3_url}{the url of the s3 bucket, e.g. s3://org-humancellatlas-upload-staging/aaaaaaaa-bbbb-cccc-dddd...}

\item{user}{the user's username e.g. `mshadbolt``}
}
\value{
a tibble of files in bucket
}
\description{
\code{list_s3_files} lists the s3 files in a given s3 bucket.
}
